overlay: fix unity-system-compositor startup
The issue is described in
https://github.com/ubports/ubuntu-touch/issues/1819: quite often, the
device fails to boot into the UI and gets stuck at the logo.

It seems that starting the USC triggers the starting of the
vendor.hwcomposer-2-2 service, but there is a race condition and USC
starts to use the display device when the Android side hasn't set it up
yet. To fix this, we manually start the service (just waiting for it
turned out not to be enough) and emit the "android" signal only once the
service is running.

My changes are those couple of lines following the "# xiaomi-violet
specific:" comment.